Handle missing 'total/up' devices metadata for RAID0 array configurations
This PR will fix the errors generated by the md check for disk instances configured with RAID0.
If an instance is setup for RAID0, the check will return a KeyError for 'total'(line #26) because the regular expression lookup fails to find a pattern matching [%d/%d] in /proc/mdstat for the instance. The C source files (links provided below) for the md driver indicate this string pattern is only found on devices configured with RAID1, RAID10, and RAID5 but not RAID0.

**Source Files:**
- RAID10: https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/tree/drivers/md/raid10.c?h=v2.6.32.5#n999
- RAID1: https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/tree/drivers/md/raid1.c?h=v2.6.32.5#n1003
- RAID5: https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git/tree/drivers/md/raid5.c#n7068